bpf: re-organize structs so we can use credentials object
